fix(git): make branch update fully update working tree
Before we were running with the safe option, which would refuse to
update any files that had uncommitted changes. However this was
calculated based on the HEAD, which we had just updated to point at the
new upstream HEAD. This means that even if your repo was clean before,
we would first update the head and then diff against it.

Fix this by manually checking that the repo is clean before changing the
HEAD. If the repo was clean we are good to force checkout the new HEAD,
and if it wasn't then we should bail early rather than updating the HEAD
but leaving the working tree unchanged.