refactor(up): switch up config toml -> yaml
Bunch of reasons for this, mainly that for several of the larger files,
toml becomes really unreadable.

Also though, there are some issues with the toml deserialisation around
nested keys and maps that make things fairly confusing.

JSON (or JSON5) would be better from an "avoiding the messiness of YAML"
PoV, but are also really hard to read at scale.

In future it might make sense to support multiple config languages, but
right now the APIs from the different serde plugins are too different to
be convenient to use here.